agri-bot/backend/requirements.txt
fastapi
uvicorn
joblib
numpy
pandas
scikit-learn
tensorflow
pillow
